mini_app/requirements.txt
GitPython==3.1.29
requests==2.28.1
Pillow==9.3.0
aiogram==2.22.2
